refactor: fold KDA kernels into FLA modules
(cherry picked from commit 43554124ba2f16b0853b64aa8a6716affc8c6ec3)
